Keep only 2 active files
